Testing-Builder refactor: move print_message to tty/log.py
